added basic snapshot test
