+++ src/App.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import App from './App';
+
+describe('<App />', () => {
+  it('renders correctly', () => {
+    const tree = renderer
+      .create(<App />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
+
